+++ extensions/calDotCom/CHANGELOG.md
@@ -3,3 +3,8 @@
 ## April 27, 2023
 
 - Get booking action: `startTime` and `endTime` data points are now of value type `date` instead of `string`.
+
+## June 30, 2023
+
+- Get booking action: add validation that bookingId and apiKey are not empty
+- Get booking action: add validation of response data to make sure we have data points